Python/redis-rq/1-QuickThumb/requirement.txt
fastapi
python-multipart
uvicorn
sqlmodel
redis
rq
Pillow